Separate connecting logic into platforms
